+++ dbdaora/exceptions.py
@@ -40,3 +40,7 @@ class InvalidGeoSpatialDataError(DBDaoraError):
 
 class CacheNotAvailableError(DBDaoraError):
     ...
+
+
+class RepositoryDataTimeoutError(DBDaoraError, TimeoutError):
+    ...