+++ docs/changelog.md
@@ -1,5 +1,9 @@
 ## CHANGELOG
 
+### UNRELEASED
+
+ - Fix repository timeout error catch
+
 ### 0.22.1 - 2020-08-28
 
  - Fix query make_key method